chatgpt_project/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
autopep8==2.3.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
distlib==0.3.9
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.5
django-allauth==65.4.1
filelock==3.17.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.63.0
packaging==24.2
pipenv==2024.4.1
platformdirs==4.3.6
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.8.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.1
